fix(manager): resolve blank screen by separating entry point
Refactored the manager page to have a dedicated entry point (index.tsx) for rendering, fixing a regression where the page would not load. The TemplateManager component is now a pure, exportable module, improving testability and resolving the issue.
