+++ src/index.css
@@ -1,10 +1,10 @@
+@import "./styles/pixel-theme.css";
+
 @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&family=IBM+Plex+Mono:wght@400;500;600;700&display=swap');
 
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-@import "./styles/pixel-theme.css";
 
 @layer base {
   :root,
