+++ src/pages/manager.html
@@ -8,9 +8,9 @@
   <link rel="stylesheet" href="/src/styles/global.css">
 </head>
 <body>
-  <div id="app">
+  <div id="root">
     <!-- App will be mounted here by React -->
   </div>
-  <script type="module" src="/src/pages/manager.tsx"></script>
+  <script type="module" src="/src/pages/index.tsx"></script>
 </body>
 </html>
